requirements.txt
Django==5.1.6
djangorestframework==3.16.0
django-cors-headers==4.7.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
opencv-python==4.10.0.84
numpy==1.26.4
requests==2.32.3
ultralytics==8.3.0
pillow==10.4.0
torch==2.5.1
torchvision==0.20.1
flask==3.0.3

